Extract the SimpleBuilder to a separate file
